+++ examples/game/src/shared/config.h
@@ -1,3 +1,5 @@
 #include <cluaupp/roblox.hpp>
 
 const int STARTING_COINS = 0;
+
+int startingCoins();